refactor workbench risk rolling route
